#1043 P5: relocate flow show cases to server_show_flow.go
Phase 5 of #1043 takes ~250 LOC out of server_show.go's
modularity-discipline violation by extracting five flow-related
ShowText case bodies (`flow-monitoring`, `flow-timeouts`,
`flow-statistics`, `sessions-top:bytes`/`sessions-top:packets`,
`flow-traceoptions`) into a new sibling file `server_show_flow.go`.

Methodology matches Phases 1-4 (#1148, #1150, #1151, #1153): semantic
relocation, no behavior change. Each case body is moved verbatim apart
from (a) `&buf` references becoming `buf` (passed-in
`*strings.Builder`) and (b) the original `if … { … } else { … }`
structures flattened into early-return form. Output is unchanged.

The `sessions-top` case takes a third `topic string` parameter (passed
as `req.Topic` from the dispatcher) so the method body no longer
references the gRPC request struct directly. The case-list itself
(both `sessions-top:bytes` and `sessions-top:packets`) still selects
the same method.

server_show.go: 3,385 → 3,142 LOC (-243).
server_show_flow.go: new, 286 LOC.

`appid` import in server_show.go is dropped (only callsite moved).
All 880+ Go tests still pass.
